+++ database/factories/SubscriptionFactory.php
@@ -50,7 +50,6 @@ class SubscriptionFactory extends Factory
             Customer::factory()->create([
                 'billable_id' => $subscription->billable_id,
                 'billable_type' => $subscription->billable_type,
-                'lemon_squeezy_id' => $subscription->lemon_squeezy_id,
             ]);
         });
     }
@@ -58,7 +57,7 @@ class SubscriptionFactory extends Factory
     /**
      * Mark the subscription as being within a trial period.
      */
-    public function trialing(DateTimeInterface $trialEndsAt = null): self
+    public function trialing(?DateTimeInterface $trialEndsAt = null): self
     {
         return $this->state([
             'status' => Subscription::STATUS_ON_TRIAL,
@@ -79,7 +78,7 @@ class SubscriptionFactory extends Factory
     /**
      * Mark the subscription as paused.
      */
-    public function paused(DateTimeInterface $resumesAt = null): self
+    public function paused(?DateTimeInterface $resumesAt = null): self
     {
         return $this->state([
             'status' => Subscription::STATUS_PAUSED,