+++ database/migrations/2023_01_16_000002_create_subscriptions_table.php
@@ -4,14 +4,12 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::create('lemon_squeezy_subscriptions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('billable_id');
-            $table->string('billable_type');
+            $table->morphs('billable');
             $table->string('type');
             $table->string('lemon_squeezy_id')->unique();
             $table->string('status');
@@ -25,8 +23,6 @@ return new class extends Migration
             $table->timestamp('renews_at')->nullable();
             $table->timestamp('ends_at')->nullable();
             $table->timestamps();
-
-            $table->index(['billable_id', 'billable_type']);
         });
     }