+++ src/Http/Controllers/WebhookController.php
@@ -5,11 +5,19 @@ namespace LemonSqueezy\Laravel\Http\Controllers;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Str;
+use LemonSqueezy\Laravel\Events\LicenseKeyCreated;
+use LemonSqueezy\Laravel\Events\LicenseKeyUpdated;
+use LemonSqueezy\Laravel\Events\OrderCreated;
+use LemonSqueezy\Laravel\Events\OrderRefunded;
 use LemonSqueezy\Laravel\Events\SubscriptionCancelled;
 use LemonSqueezy\Laravel\Events\SubscriptionCreated;
 use LemonSqueezy\Laravel\Events\SubscriptionExpired;
 use LemonSqueezy\Laravel\Events\SubscriptionPaused;
+use LemonSqueezy\Laravel\Events\SubscriptionPaymentFailed;
+use LemonSqueezy\Laravel\Events\SubscriptionPaymentRecovered;
+use LemonSqueezy\Laravel\Events\SubscriptionPaymentSuccess;
 use LemonSqueezy\Laravel\Events\SubscriptionResumed;
 use LemonSqueezy\Laravel\Events\SubscriptionUnpaused;
 use LemonSqueezy\Laravel\Events\SubscriptionUpdated;
@@ -18,6 +26,7 @@ use LemonSqueezy\Laravel\Events\WebhookReceived;
 use LemonSqueezy\Laravel\Exceptions\InvalidCustomPayload;
 use LemonSqueezy\Laravel\Http\Middleware\VerifyWebhookSignature;
 use LemonSqueezy\Laravel\LemonSqueezy;
+use LemonSqueezy\Laravel\Order;
 use LemonSqueezy\Laravel\Subscription;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -44,7 +53,7 @@ final class WebhookController extends Controller
             return new Response('Webhook received but no event name was found.');
         }
 
-        $method = 'handle'.Str::studly($payload['meta']['event_name']);
+        $method = 'handle' . Str::studly($payload['meta']['event_name']);
 
         WebhookReceived::dispatch($payload);
 
@@ -63,31 +72,71 @@ final class WebhookController extends Controller
         return new Response('Webhook received but no handler found.');
     }
 
-    /**
-     * @throws InvalidCustomPayload
-     */
+    public function handleOrderCreated(array $payload): void
+    {
+        $billable = $this->resolveBillable($payload);
+
+        // Todo v2: Remove this check
+        if (Schema::hasTable((new LemonSqueezy::$orderModel())->getTable())) {
+            $attributes = $payload['data']['attributes'];
+
+            $order = $billable->orders()->create([
+                'lemon_squeezy_id' => $payload['data']['id'],
+                'customer_id' => $attributes['customer_id'],
+                'product_id' => (string) $attributes['first_order_item']['product_id'],
+                'variant_id' => (string) $attributes['first_order_item']['variant_id'],
+                'identifier' => $attributes['identifier'],
+                'order_number' => $attributes['order_number'],
+                'currency' => $attributes['currency'],
+                'subtotal' => $attributes['subtotal'],
+                'discount_total' => $attributes['discount_total'],
+                'tax' => $attributes['tax'],
+                'total' => $attributes['total'],
+                'tax_name' => $attributes['tax_name'],
+                'status' => $attributes['status'],
+                'receipt_url' => $attributes['urls']['receipt'] ?? null,
+                'refunded' => $attributes['refunded'],
+                'refunded_at' => $attributes['refunded_at'] ? Carbon::make($attributes['refunded_at']) : null,
+                'ordered_at' => Carbon::make($attributes['created_at']),
+            ]);
+        } else {
+            $order = null;
+        }
+
+        OrderCreated::dispatch($billable, $order, $payload);
+    }
+
+    public function handleOrderRefunded(array $payload): void
+    {
+        $billable = $this->resolveBillable($payload);
+
+        // Todo v2: Remove this check
+        if (Schema::hasTable((new LemonSqueezy::$orderModel())->getTable())) {
+            if (! $order = $this->findOrder($payload['data']['id'])) {
+                return;
+            }
+
+            $order = $order->sync($payload['data']['attributes']);
+        } else {
+            $order = null;
+        }
+
+        OrderRefunded::dispatch($billable, $order, $payload);
+    }
+
     public function handleSubscriptionCreated(array $payload): void
     {
         $custom = $payload['meta']['custom_data'] ?? null;
-
-        if (! isset($custom) || ! is_array($custom) || ! isset($custom['billable_id'], $custom['billable_type'])) {
-            throw new InvalidCustomPayload;
-        }
-
         $attributes = $payload['data']['attributes'];
 
-        $billable = $this->findOrCreateCustomer(
-            $custom['billable_id'],
-            (string) $custom['billable_type'],
-            (string) $attributes['customer_id']
-        );
+        $billable = $this->resolveBillable($payload);
 
         $subscription = $billable->subscriptions()->create([
             'type' => $custom['subscription_type'] ?? Subscription::DEFAULT_TYPE,
             'lemon_squeezy_id' => $payload['data']['id'],
             'status' => $attributes['status'],
-            'product_id' => $attributes['product_id'],
-            'variant_id' => $attributes['variant_id'],
+            'product_id' => (string) $attributes['product_id'],
+            'variant_id' => (string) $attributes['variant_id'],
             'card_brand' => $attributes['card_brand'] ?? null,
             'card_last_four' => $attributes['card_last_four'] ?? null,
             'trial_ends_at' => $attributes['trial_ends_at'] ? Carbon::make($attributes['trial_ends_at']) : null,
@@ -98,6 +147,11 @@ final class WebhookController extends Controller
         // Terminate the billable's generic trial at the model level if it exists...
         if (! is_null($billable->customer->trial_ends_at)) {
             $billable->customer->update(['trial_ends_at' => null]);
+        }
+
+        // Set the billable's lemon squeezy id if it was on generic trial at the model level
+        if (is_null($billable->customer->lemon_squeezy_id)) {
+            $billable->customer->update(['lemon_squeezy_id' => $attributes['customer_id']]);
         }
 
         SubscriptionCreated::dispatch($billable, $subscription, $payload);
@@ -111,7 +165,9 @@ final class WebhookController extends Controller
 
         $subscription = $subscription->sync($payload['data']['attributes']);
 
-        SubscriptionUpdated::dispatch($subscription->billable, $subscription, $payload);
+        if ($subscription->billable) {
+            SubscriptionUpdated::dispatch($subscription->billable, $subscription, $payload);
+        }
     }
 
     private function handleSubscriptionCancelled(array $payload): void
@@ -122,7 +178,9 @@ final class WebhookController extends Controller
 
         $subscription = $subscription->sync($payload['data']['attributes']);
 
-        SubscriptionCancelled::dispatch($subscription->billable, $subscription, $payload);
+        if ($subscription->billable) {
+            SubscriptionCancelled::dispatch($subscription->billable, $subscription, $payload);
+        }
     }
 
     private function handleSubscriptionResumed(array $payload): void
@@ -144,7 +202,9 @@ final class WebhookController extends Controller
 
         $subscription = $subscription->sync($payload['data']['attributes']);
 
-        SubscriptionExpired::dispatch($subscription->billable, $subscription, $payload);
+        if ($subscription->billable) {
+            SubscriptionExpired::dispatch($subscription->billable, $subscription, $payload);
+        }
     }
 
     private function handleSubscriptionPaused(array $payload): void
@@ -169,6 +229,70 @@ final class WebhookController extends Controller
         SubscriptionUnpaused::dispatch($subscription->billable, $subscription, $payload);
     }
 
+    private function handleSubscriptionPaymentSuccess(array $payload): void
+    {
+        if (
+            ($subscription = $this->findSubscription($payload['data']['attributes']['subscription_id'])) &&
+            $subscription->billable
+        ) {
+            SubscriptionPaymentSuccess::dispatch($subscription->billable, $subscription, $payload);
+        }
+    }
+
+    private function handleSubscriptionPaymentFailed(array $payload): void
+    {
+        if (
+            ($subscription = $this->findSubscription($payload['data']['attributes']['subscription_id'])) &&
+            $subscription->billable
+        ) {
+            SubscriptionPaymentFailed::dispatch($subscription->billable, $subscription, $payload);
+        }
+    }
+
+    private function handleSubscriptionPaymentRecovered(array $payload): void
+    {
+        if (
+            ($subscription = $this->findSubscription($payload['data']['attributes']['subscription_id'])) &&
+            $subscription->billable
+        ) {
+            SubscriptionPaymentRecovered::dispatch($subscription->billable, $subscription, $payload);
+        }
+    }
+
+    private function handleLicenseKeyCreated(array $payload): void
+    {
+        $billable = $this->resolveBillable($payload);
+
+        LicenseKeyCreated::dispatch($billable, $payload);
+    }
+
+    private function handleLicenseKeyUpdated(array $payload): void
+    {
+        $billable = $this->resolveBillable($payload);
+
+        LicenseKeyUpdated::dispatch($billable, $payload);
+    }
+
+    /**
+     * @return \LemonSqueezy\Laravel\Billable
+     *
+     * @throws InvalidCustomPayload
+     */
+    private function resolveBillable(array $payload)
+    {
+        $custom = $payload['meta']['custom_data'] ?? null;
+
+        if (! isset($custom) || ! is_array($custom) || ! isset($custom['billable_id'], $custom['billable_type'])) {
+            throw new InvalidCustomPayload();
+        }
+
+        return $this->findOrCreateCustomer(
+            $custom['billable_id'],
+            (string) $custom['billable_type'],
+            (string) $payload['data']['attributes']['customer_id'],
+        );
+    }
+
     /**
      * @return \LemonSqueezy\Laravel\Billable
      */
@@ -185,5 +309,10 @@ final class WebhookController extends Controller
     private function findSubscription(string $subscriptionId): ?Subscription
     {
         return LemonSqueezy::$subscriptionModel::firstWhere('lemon_squeezy_id', $subscriptionId);
+    }
+
+    private function findOrder(string $orderId): ?Order
+    {
+        return LemonSqueezy::$orderModel::firstWhere('lemon_squeezy_id', $orderId);
     }
 }