+++ src/LemonSqueezyServiceProvider.php
@@ -5,6 +5,9 @@ namespace LemonSqueezy\Laravel;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
+use LemonSqueezy\Laravel\Console\ListenCommand;
+use LemonSqueezy\Laravel\Console\ListLicensesCommand;
+use LemonSqueezy\Laravel\Console\ListProductsCommand;
 use LemonSqueezy\Laravel\Http\Controllers\WebhookController;
 
 class LemonSqueezyServiceProvider extends ServiceProvider
@@ -12,7 +15,8 @@ class LemonSqueezyServiceProvider extends ServiceProvider
     public function register(): void
     {
         $this->mergeConfigFrom(
-            __DIR__.'/../config/lemon-squeezy.php', 'lemon-squeezy'
+            __DIR__ . '/../config/lemon-squeezy.php',
+            'lemon-squeezy',
         );
     }
 
@@ -24,6 +28,7 @@ class LemonSqueezyServiceProvider extends ServiceProvider
         $this->bootPublishing();
         $this->bootDirectives();
         $this->bootComponents();
+        $this->bootCommands();
     }
 
     protected function bootRoutes(): void
@@ -40,13 +45,13 @@ class LemonSqueezyServiceProvider extends ServiceProvider
 
     protected function bootResources(): void
     {
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'lemon-squeezy');
+        $this->loadViewsFrom(__DIR__ . '/../resources/views', 'lemon-squeezy');
     }
 
     protected function bootMigrations(): void
     {
         if (LemonSqueezy::$runsMigrations && $this->app->runningInConsole()) {
-            $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+            $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
         }
     }
 
@@ -54,15 +59,15 @@ class LemonSqueezyServiceProvider extends ServiceProvider
     {
         if ($this->app->runningInConsole()) {
             $this->publishes([
-                __DIR__.'/../config/lemon-squeezy.php' => $this->app->configPath('lemon-squeezy.php'),
+                __DIR__ . '/../config/lemon-squeezy.php' => $this->app->configPath('lemon-squeezy.php'),
             ], 'lemon-squeezy-config');
 
             $this->publishes([
-                __DIR__.'/../database/migrations' => $this->app->databasePath('migrations'),
+                __DIR__ . '/../database/migrations' => $this->app->databasePath('migrations'),
             ], 'lemon-squeezy-migrations');
 
             $this->publishes([
-                __DIR__.'/../resources/views' => $this->app->resourcePath('views/vendor/lemon-squeezy'),
+                __DIR__ . '/../resources/views' => $this->app->resourcePath('views/vendor/lemon-squeezy'),
             ], 'lemon-squeezy-views');
         }
     }
@@ -77,5 +82,16 @@ class LemonSqueezyServiceProvider extends ServiceProvider
     protected function bootComponents(): void
     {
         Blade::component('lemon-squeezy::components.button', 'lemon-button');
+    }
+
+    protected function bootCommands(): void
+    {
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                ListenCommand::class,
+                ListLicensesCommand::class,
+                ListProductsCommand::class,
+            ]);
+        }
     }
 }