+++ CHANGELOG.md
@@ -1,6 +1,143 @@
 # Release Notes
 
-## [Unreleased](https://github.com/lmsqueezy/laravel/compare/0.2.0...main)
+## [Unreleased](https://github.com/lmsqueezy/laravel/compare/1.8.2...main)
+
+## [1.8.2](https://github.com/lmsqueezy/laravel/compare/1.8.1...1.8.2) - 2025-02-13
+
+### What's Changed
+
+* Fix deprecated variable syntax by [@virgildotcodes](https://github.com/virgildotcodes) in https://github.com/lmsqueezy/laravel/pull/119
+* Creating pre-release branch by [@JustSteveKing](https://github.com/JustSteveKing) in https://github.com/lmsqueezy/laravel/pull/120
+* Pre release by [@JustSteveKing](https://github.com/JustSteveKing) in https://github.com/lmsqueezy/laravel/pull/121
+* Preparing release by [@JustSteveKing](https://github.com/JustSteveKing) in https://github.com/lmsqueezy/laravel/pull/122
+
+### New Contributors
+
+* [@virgildotcodes](https://github.com/virgildotcodes) made their first contribution in https://github.com/lmsqueezy/laravel/pull/119
+* [@JustSteveKing](https://github.com/JustSteveKing) made their first contribution in https://github.com/lmsqueezy/laravel/pull/120
+
+**Full Changelog**: https://github.com/lmsqueezy/laravel/compare/1.8.1...1.8.2
+
+## [1.8.1](https://github.com/lmsqueezy/laravel/compare/1.8.0...1.8.1) - 2025-01-13
+
+* Fixes valid state of subscription by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/116
+
+## [1.8.0](https://github.com/lmsqueezy/laravel/compare/1.7.0...1.8.0) - 2025-01-13
+
+* Create list licenses console command by [@paulinevos](https://github.com/paulinevos) in https://github.com/lmsqueezy/laravel/pull/115
+
+## [1.7.0](https://github.com/lmsqueezy/laravel/compare/1.6.3...1.7.0) - 2024-10-16
+
+* PHP 8.4 support by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/109
+
+## [1.6.3](https://github.com/lmsqueezy/laravel/compare/1.6.2...1.6.3) - 2024-10-07
+
+* Add support for prompts 0.2 and 0.3 by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/107
+
+## [1.6.2](https://github.com/lmsqueezy/laravel/compare/1.6.1...1.6.2) - 2024-07-23
+
+* fix: clean up annoying empty lines and timer for expose listen in terminal by [@boudydegeer](https://github.com/boudydegeer) in https://github.com/lmsqueezy/laravel/pull/100
+
+## [1.6.1](https://github.com/lmsqueezy/laravel/compare/1.6.0...1.6.1) - 2024-07-23
+
+* fix: Filter correct products for store in ListProductsCommand by [@boudydegeer](https://github.com/boudydegeer) in https://github.com/lmsqueezy/laravel/pull/99
+
+## [1.6.0](https://github.com/lmsqueezy/laravel/compare/1.5.5...1.6.0) - 2024-07-03
+
+* Add command to list products and variants by [@heyjorgedev](https://github.com/heyjorgedev) in https://github.com/lmsqueezy/laravel/pull/80
+
+## [1.5.5](https://github.com/lmsqueezy/laravel/compare/1.5.4...1.5.5) - 2024-07-01
+
+* Limit Listen command on Windows by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/97
+
+## [1.5.4](https://github.com/lmsqueezy/laravel/compare/1.5.3...1.5.4) - 2024-05-27
+
+* Prevent firing of events when billable is deleted by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/91
+
+## [1.5.3](https://github.com/lmsqueezy/laravel/compare/1.5.2...1.5.3) - 2024-05-06
+
+* Replace carbon return type with carboninterface for trialEndsAt by [@dimitri-koenig](https://github.com/dimitri-koenig) in https://github.com/lmsqueezy/laravel/pull/87
+
+## [1.5.2](https://github.com/lmsqueezy/laravel/compare/1.5.1...1.5.2) - 2024-03-26
+
+* Fix Tests on Laravel 11 and exit codes by [@heyjorgedev](https://github.com/heyjorgedev) in https://github.com/lmsqueezy/laravel/pull/79
+* Cast variant_id before saving by [@matthiasweiss](https://github.com/matthiasweiss) in https://github.com/lmsqueezy/laravel/pull/81
+
+## [1.5.1](https://github.com/lmsqueezy/laravel/compare/1.5.0...1.5.1) - 2024-03-21
+
+* Fix listen command with webhook secret by [@DanielHudson](https://github.com/DanielHudson) in https://github.com/lmsqueezy/laravel/pull/78
+
+## [1.5.0](https://github.com/lmsqueezy/laravel/compare/1.4.0...1.5.0) - 2024-03-08
+
+* Laravel v11 by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/75
+
+## [1.4.0](https://github.com/lmsqueezy/laravel/compare/1.3.3...1.4.0) - 2023-12-12
+
+* `artisan lmsqueezy:listen` by [@DanielHudson](https://github.com/DanielHudson) in https://github.com/lmsqueezy/laravel/pull/54
+
+## [1.3.3](https://github.com/lmsqueezy/laravel/compare/1.3.2...1.3.3) - 2023-12-12
+
+- Fix missing Illuminate Builder import in Order model by [@rvetrsek](https://github.com/rvetrsek) in https://github.com/lmsqueezy/laravel/pull/68
+
+## [1.3.2](https://github.com/lmsqueezy/laravel/compare/1.3.1...1.3.2) - 2023-11-20
+
+- Add indexes by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/63
+
+## [1.3.1](https://github.com/lmsqueezy/laravel/compare/1.3.0...1.3.1) - 2023-11-20
+
+- Fixes for orders by [@calebporzio](https://github.com/calebporzio) in https://github.com/lmsqueezy/laravel/pull/62
+
+## [1.3.0](https://github.com/lmsqueezy/laravel/compare/1.2.4...1.3.0) - 2023-11-19
+
+- Implement Orders by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/60
+
+## [1.2.4](https://github.com/lmsqueezy/laravel/compare/1.2.3...1.2.4) - 2023-11-02
+
+- Fix missing lemon_squeezy_id if billable was on generic trial by [@pdziewa](https://github.com/pdziewa) in https://github.com/lmsqueezy/laravel/pull/52
+
+## [1.2.3](https://github.com/lmsqueezy/laravel/compare/1.2.2...1.2.3) - 2023-10-10
+
+- Fix nullable lemon_squeezy_id on customers table by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/50
+
+## [1.2.2](https://github.com/lmsqueezy/laravel/compare/1.2.1...1.2.2) - 2023-10-10
+
+- PHP 8.3 support by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/48
+
+## [1.2.1](https://github.com/lmsqueezy/laravel/compare/1.2.0...1.2.1) - 2023-10-05
+
+- Fix issues with customer record by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/47
+
+## [1.2.0](https://github.com/lmsqueezy/laravel/compare/1.1.2...1.2.0) - 2023-10-04
+
+- Support for customer portals by [@driesvints](https://github.com/driesvints) in https://github.com/lmsqueezy/laravel/pull/46
+- Use `hash_equals` when validating webhook signature by [@thinkverse](https://github.com/thinkverse) in https://github.com/lmsqueezy/laravel/pull/42
+
+## [1.1.2](https://github.com/lmsqueezy/laravel/compare/1.1.1...1.1.2) - 2023-08-10
+
+- Adding Event and Webhook handler to get license updates by [@abishekrsrikaanth](https://github.com/abishekrsrikaanth) in https://github.com/lmsqueezy/laravel/pull/41
+
+## [1.1.1](https://github.com/lmsqueezy/laravel/compare/1.1.0...1.1.1) - 2023-08-06
+
+- Dispatching the correct event name when a license is created by [@abishekrsrikaanth](https://github.com/abishekrsrikaanth) in https://github.com/lmsqueezy/laravel/pull/35
+
+## [1.1.0](https://github.com/lmsqueezy/laravel/compare/1.0.1...1.1.0) - 2023-07-18
+
+- Add support for custom name, description and receipt_thank_you_note at checkout by [@UrosCodes](https://github.com/UrosCodes) in https://github.com/lmsqueezy/laravel/pull/34
+
+## [1.0.1](https://github.com/lmsqueezy/laravel/compare/1.0.0...1.0.1) - 2023-07-15
+
+- Fix can't find Subscription from SubscriptionPayment by [@ycs77](https://github.com/ycs77) in https://github.com/lmsqueezy/laravel/pull/33
+
+## [1.0.0](https://github.com/lmsqueezy/laravel/compare/0.2.1...1.0.0) - 2023-06-16
+
+- First stable release
+- Add support for custom price at checkout by @mraheelkhan in https://github.com/lmsqueezy/laravel/pull/28
+- Add remaining events by @driesvints in https://github.com/lmsqueezy/laravel/pull/30
+
+## [0.2.1](https://github.com/lmsqueezy/laravel/compare/0.2.0...0.2.1) - 2023-05-20
+
+- Add "down" method for migrations by @Flatroy in https://github.com/lmsqueezy/laravel/pull/19
+- Add endTrial method and document billing anchor by @driesvints in https://github.com/lmsqueezy/laravel/commit/2d63056363fd345ef8971d2d1651f628ef7599f1
 
 ## [0.2.0](https://github.com/lmsqueezy/laravel/compare/0.1.0...0.2.0) - 2023-05-08